Struct plus1::Comment

pub struct Comment {
    pub in_reply_to: Vec<CommentInReplyTo>,
    pub kind: Option<String>,
    pub plusoners: Option<CommentPlusoners>,
    pub object: Option<CommentObject>,
    pub updated: Option<String>,
    pub actor: Option<CommentActor>,
    pub verb: Option<String>,
    pub etag: Option<String>,
    pub published: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list comments (none)

get comments (response)

Fields

in_reply_to

The activity this comment replied to.

kind

Identifies this resource as a comment. Value: "plus#comment".

plusoners

People who +1'd this comment.

object

The object of this comment.

updated

The time at which this comment was last updated. Formatted as an RFC 3339 timestamp.

actor

The person who posted this comment.

verb

This comment's verb, indicating what action was performed. Possible values are: - "post" - Publish content to the stream.

etag

ETag of this response for caching purposes.

published

The time at which this comment was initially published. Formatted as an RFC 3339 timestamp.

id

The ID of this comment.

self_link

Link to this comment resource.